Fixed Arthimetic words
Fixed issues with double conversion.

+++ src/edu/mccc/cos210/ds/fp/javaforth/words/Minus.java
@@ -11,9 +11,9 @@ public class Minus extends ForthWordBase {
 	@Override
 	public void execute(IStack<Object> stack, ForthDictionary dict) {
 		try {
-			int n1 = (int) stack.pop();
-			int n2 = (int) stack.pop();
-			stack.push(Integer.valueOf((n1-n2)));
+			int n1 = ((Double) stack.pop()).intValue();
+			int n2 = ((Double) stack.pop()).intValue();
+			stack.push(Double.valueOf((n1-n2)));
 		}catch(Exception e) {
 			throw new RuntimeException("Error on word -");
 		}
